extract ExtensionPromptFileService (#315815)
